perf: cache esm modules by content hash
The cache option has been removed as caching is now enabled by default and it is based on each
module's content. When the module's content change, the cache is automatically invalidated and the
updated version is served (and cached again).

BREAKING CHANGE: Requesting a module located outside the `root` and `nodeModulesRoot` absolute paths
is not supported anymore. This behavior was removed because it represented a security issue.